+++ packages/query-graphql/__tests__/types/delete-one-input.type.spec.ts
@@ -2,16 +2,16 @@ import 'reflect-metadata';
 import * as typeGraphql from 'type-graphql';
 import { plainToClass } from 'class-transformer';
 import { validateSync } from 'class-validator';
-import { DeleteOneArgsType } from '../../src';
+import { DeleteOneInputType } from '../../src';
 
-describe('DeleteOneArgsType', (): void => {
-  const argsTypeSpy = jest.spyOn(typeGraphql, 'ArgsType');
+describe('DeleteOneInputType', (): void => {
+  const inputTypeSpy = jest.spyOn(typeGraphql, 'InputType');
   const fieldSpy = jest.spyOn(typeGraphql, 'Field');
 
   it('should create an args type with the field as the type', () => {
-    DeleteOneArgsType();
-    expect(argsTypeSpy).toBeCalledWith();
-    expect(argsTypeSpy).toBeCalledTimes(1);
+    DeleteOneInputType();
+    expect(inputTypeSpy).toBeCalledWith();
+    expect(inputTypeSpy).toBeCalledTimes(1);
     expect(fieldSpy).toBeCalledTimes(1);
     expect(fieldSpy.mock.calls[0]![0]!()).toEqual(typeGraphql.ID);
   });
@@ -19,31 +19,31 @@ describe('DeleteOneArgsType', (): void => {
   describe('validation', () => {
     it('should validate the id is defined', () => {
       const input = {};
-      const it = plainToClass(DeleteOneArgsType(), input);
+      const it = plainToClass(DeleteOneInputType(), input);
       const errors = validateSync(it);
       expect(errors).toEqual([
         {
           children: [],
           constraints: {
-            isNotEmpty: 'input should not be empty',
+            isNotEmpty: 'id should not be empty',
           },
-          property: 'input',
+          property: 'id',
           target: input,
         },
       ]);
     });
 
     it('should validate the id is not empty', () => {
-      const input = { input: '' };
-      const it = plainToClass(DeleteOneArgsType(), input);
+      const input = { id: '' };
+      const it = plainToClass(DeleteOneInputType(), input);
       const errors = validateSync(it);
       expect(errors).toEqual([
         {
           children: [],
           constraints: {
-            isNotEmpty: 'input should not be empty',
+            isNotEmpty: 'id should not be empty',
           },
-          property: 'input',
+          property: 'id',
           target: input,
           value: '',
         },