+++ packages/query-graphql/__tests__/types/update-one-input.type.spec.ts
@@ -2,36 +2,31 @@ import 'reflect-metadata';
 import * as typeGraphql from 'type-graphql';
 import { plainToClass } from 'class-transformer';
 import { validateSync, MinLength } from 'class-validator';
-import { UpdateOneArgsType } from '../../src';
+import { ObjectType } from 'type-graphql';
+import { UpdateOneInputType } from '../../src';
 
 describe('UpdateOneInputType', (): void => {
-  const argsTypeSpy = jest.spyOn(typeGraphql, 'ArgsType');
+  const inputType = jest.spyOn(typeGraphql, 'InputType');
   const fieldSpy = jest.spyOn(typeGraphql, 'Field');
 
+  @ObjectType()
   class FakeType {
     @MinLength(5)
     name!: string;
   }
   it('should create an args type with the field as the type', () => {
-    UpdateOneArgsType(FakeType);
-    expect(argsTypeSpy).toBeCalledWith();
-    expect(argsTypeSpy).toBeCalledTimes(1);
+    UpdateOneInputType(FakeType, FakeType);
+    expect(inputType).toBeCalledTimes(1);
+    expect(inputType).toBeCalledWith(`UpdateOneFakeTypeInput`);
     expect(fieldSpy).toBeCalledTimes(2);
     expect(fieldSpy.mock.calls[0]![0]!()).toEqual(typeGraphql.ID);
     expect(fieldSpy.mock.calls[1]![0]!()).toEqual(FakeType);
   });
 
-  it('should return the input when accessing the update field', () => {
-    const Type = UpdateOneArgsType(FakeType);
-    const input = { id: 1, input: {} };
-    const it = plainToClass(Type, input);
-    expect(it.input).toEqual(input.input);
-  });
-
   describe('validation', () => {
     it('should validate id is defined is not empty', () => {
-      const Type = UpdateOneArgsType(FakeType);
-      const input = { input: { name: 'hello world' } };
+      const Type = UpdateOneInputType(FakeType, FakeType);
+      const input = { update: { name: 'hello world' } };
       const it = plainToClass(Type, input);
       const errors = validateSync(it);
       expect(errors).toEqual([
@@ -47,8 +42,8 @@ describe('UpdateOneInputType', (): void => {
     });
 
     it('should validate id is not empty is defined is not empty', () => {
-      const Type = UpdateOneArgsType(FakeType);
-      const input = { id: '', input: { name: 'hello world' } };
+      const Type = UpdateOneInputType(FakeType, FakeType);
+      const input = { id: '', update: { name: 'hello world' } };
       const it = plainToClass(Type, input);
       const errors = validateSync(it);
       expect(errors).toEqual([
@@ -65,8 +60,8 @@ describe('UpdateOneInputType', (): void => {
     });
 
     it('should validate the update input', () => {
-      const Type = UpdateOneArgsType(FakeType);
-      const input = { id: 'id-1', input: {} };
+      const Type = UpdateOneInputType(FakeType, FakeType);
+      const input = { id: 'id-1', update: {} };
       const it = plainToClass(Type, input);
       const errors = validateSync(it);
       expect(errors).toEqual([
@@ -81,9 +76,9 @@ describe('UpdateOneInputType', (): void => {
               target: {},
             },
           ],
-          property: 'input',
+          property: 'update',
           target: it,
-          value: it.input,
+          value: it.update,
         },
       ]);
     });